ece651-sp22-zg79-battleship/app/src/main/java/edu/duke/zg79/battleship/NoCollisionRuleChecker.java
package edu.duke.zg79.battleship;

/**
 * This class check the rule that the ship does not 
 * collide with anything else on theBoard
 */
public class NoCollisionRuleChecker<T> extends PlacementRuleChecker<T> {

  /**
   * The next placement rule checker
   */
  public NoCollisionRuleChecker(PlacementRuleChecker<T> next) {
    super(next);
  }

  /**
   * This check the placement of ship does not collide with other ships 
   */
  @Override
  protected String checkMyRule(Ship<T> theShip, Board<T> theBoard) {
    Iterable<Coordinate> shipSet = theShip.getCoordinates();
    for (Coordinate c: shipSet) {
      if (theBoard.whatIsAtForSelf(c) != null) {
        return "That placement is invalid: the ship overlaps another ship.";
      }
    }
    return null;
  }
}
